Move check private key sync to account ctrl
